requirements.txt
boto3==1.40.13
botocore==1.40.13
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
idna==3.10
jmespath==1.0.1
MIDIUtil==1.2.1
mido==1.3.3
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
py4j==0.10.9.9
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.13.1
six==1.17.0
spotipy==2.25.1
tweepy==4.16.0
urllib3==2.3.0
